+++ src/main/java/controllers/imageViewer/ImageFilters.java
@@ -1,4 +1,4 @@
-package controllers;
+package controllers.imageViewer;
 
 /**
  * @author (created on 12/5/2017).
